PlayerMovement Script
Made a dummy scene to test movement, it's not perfect and still missing throw, but need carrot sprite for that. Moving left and right, jump and double jump all working ok. I think there might be some problems with the animation, need you guys to take a look with me.